[refactor/#179] questionnaire test 수정
